Keep communicating with server
In order to implement the functionality of a live chatroom, we need to
keep commnicating with the server. Moreover, we need to be able to send
message to the server and receive the message from other clients. The
server will broadcast the message it gets from other clients, all we
need to do is to read the message sent by the server.

- Use "select" to monitor we need to do "read" or "write"

- Use a infinite loop to keep communicating with the server until
disconnection