Fix a little bug
Before this commit, when a message is sent from clients, the message
is duplicate and shown on the server screen about three times, which
we didn't mean to do that. I found that is because in the for loop of
the write operation, sometimes j is not a socket number but we still do
write operation on it.

- Add a new condition to the first if statement in the for loop to
check if j is in the file description set