Use struct "message" as package for commnication
Before this change, I used a string "buf" as message. However, we
can't identify which kind of message it is, and we can't know the
username of the client who sent this message unless we read twice.
In this case, I reprogrammed almost all the implementation in the
loop by replacing a character array "buf" by a struct "message",
which contains username, message type, and message.

- From clients to server, there are two types of message: normal message
and set username. (username setting is the first message sent from
each client after it's connected to the server)

- From the server to clients, there are three types of message: normal
message from other clients, the notification of disconnection of other
clients, and the notification of connection of other clients